Added user authorization in CRUD features

+++ client/src/pages/DeleteProblem.jsx
@@ -1,18 +1,34 @@
 import axios from "axios";
 import React from "react";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import "../css/DeleteProblem.css";
 const DeleteProblem = (props) => {
   const { _id, title, statement, difficulty } = props.obj;
+  const navigate = useNavigate();
   const deleteProblem = async () => {
     try {
-      const response = await axios.delete(
-        `http://localhost:5000/problems/${_id}`
+      const response = await axios.post(
+        "http://localhost:5000",
+        {},
+        { withCredentials: true }
       );
-      console.log("Problem gets deleted");
-      window.location.reload();
+      // console.log(response);
+      if (!response.data.success) {
+        console.log("token not found Please login first");
+        navigate("/login");
+      } else {
+        try {
+          const response = await axios.delete(
+            `http://localhost:5000/problems/${_id}`
+          );
+          console.log("Problem gets deleted");
+          window.location.reload();
+        } catch (error) {
+          console.log(error.response.data.message);
+        }
+      }
     } catch (error) {
-      console.log(error.response.data.message);
+      console.log("Some error occurred");
     }
   };
   return (
